Move widgets into separate packages
