+++ WordPress/src/main/java/org/wordpress/android/modules/AppComponent.java
@@ -137,12 +137,12 @@ import org.wordpress.android.ui.stats.StatsWidgetConfigureAdapter;
 import org.wordpress.android.ui.stats.StatsWidgetProvider;
 import org.wordpress.android.ui.stats.refresh.StatsActivity;
 import org.wordpress.android.ui.stats.refresh.StatsModule;
-import org.wordpress.android.ui.stats.refresh.lists.widget.AllTimeWidgetListProvider;
-import org.wordpress.android.ui.stats.refresh.lists.widget.StatsAllTimeWidget;
-import org.wordpress.android.ui.stats.refresh.lists.widget.StatsTodayWidget;
-import org.wordpress.android.ui.stats.refresh.lists.widget.StatsViewsWidget;
-import org.wordpress.android.ui.stats.refresh.lists.widget.TodayWidgetListProvider;
-import org.wordpress.android.ui.stats.refresh.lists.widget.ViewsWidgetListProvider;
+import org.wordpress.android.ui.stats.refresh.lists.widget.alltime.AllTimeWidgetListProvider;
+import org.wordpress.android.ui.stats.refresh.lists.widget.alltime.StatsAllTimeWidget;
+import org.wordpress.android.ui.stats.refresh.lists.widget.today.StatsTodayWidget;
+import org.wordpress.android.ui.stats.refresh.lists.widget.views.StatsViewsWidget;
+import org.wordpress.android.ui.stats.refresh.lists.widget.today.TodayWidgetListProvider;
+import org.wordpress.android.ui.stats.refresh.lists.widget.views.ViewsWidgetListProvider;
 import org.wordpress.android.ui.stats.service.StatsServiceLogic;
 import org.wordpress.android.ui.stockmedia.StockMediaPickerActivity;
 import org.wordpress.android.ui.suggestion.adapters.SuggestionAdapter;
