+++ WordPress/src/main/java/org/wordpress/android/modules/ApplicationModule.java
@@ -16,9 +16,9 @@ import org.wordpress.android.ui.stats.refresh.StatsViewAllFragment;
 import org.wordpress.android.ui.stats.refresh.lists.StatsListFragment;
 import org.wordpress.android.ui.stats.refresh.lists.detail.StatsDetailFragment;
 import org.wordpress.android.ui.stats.refresh.lists.sections.insights.management.InsightsManagementFragment;
-import org.wordpress.android.ui.stats.refresh.lists.widget.StatsWidgetSiteSelectionDialogFragment;
-import org.wordpress.android.ui.stats.refresh.lists.widget.StatsWidgetConfigureFragment;
-import org.wordpress.android.ui.stats.refresh.lists.widget.ColorSelectionDialogFragment;
+import org.wordpress.android.ui.stats.refresh.lists.widget.configuration.StatsWidgetSiteSelectionDialogFragment;
+import org.wordpress.android.ui.stats.refresh.lists.widget.configuration.StatsWidgetConfigureFragment;
+import org.wordpress.android.ui.stats.refresh.lists.widget.configuration.ColorSelectionDialogFragment;
 import org.wordpress.android.util.wizard.WizardManager;
 import org.wordpress.android.viewmodel.helpers.ConnectionStatus;
 import org.wordpress.android.viewmodel.helpers.ConnectionStatusLiveData;
