+++ WordPress/src/main/java/org/wordpress/android/ui/stats/refresh/lists/widget/WidgetService.kt
@@ -2,9 +2,13 @@ package org.wordpress.android.ui.stats.refresh.lists.widget
 
 import android.content.Intent
 import android.widget.RemoteViewsService
-import org.wordpress.android.ui.stats.refresh.lists.widget.StatsWidgetConfigureFragment.ViewType.ALL_TIME_VIEWS
-import org.wordpress.android.ui.stats.refresh.lists.widget.StatsWidgetConfigureFragment.ViewType.TODAY_VIEWS
-import org.wordpress.android.ui.stats.refresh.lists.widget.StatsWidgetConfigureFragment.ViewType.WEEK_VIEWS
+import org.wordpress.android.ui.stats.refresh.lists.widget.configuration.StatsWidgetConfigureFragment.ViewType.ALL_TIME_VIEWS
+import org.wordpress.android.ui.stats.refresh.lists.widget.configuration.StatsWidgetConfigureFragment.ViewType.TODAY_VIEWS
+import org.wordpress.android.ui.stats.refresh.lists.widget.configuration.StatsWidgetConfigureFragment.ViewType.WEEK_VIEWS
+import org.wordpress.android.ui.stats.refresh.lists.widget.alltime.AllTimeWidgetListProvider
+import org.wordpress.android.ui.stats.refresh.lists.widget.today.TodayWidgetListProvider
+import org.wordpress.android.ui.stats.refresh.lists.widget.utils.getViewType
+import org.wordpress.android.ui.stats.refresh.lists.widget.views.ViewsWidgetListProvider
 
 class WidgetService : RemoteViewsService() {
     override fun onGetViewFactory(intent: Intent): RemoteViewsFactory {
