+++ WordPress/src/main/java/org/wordpress/android/ui/stats/refresh/lists/widget/today/StatsTodayWidget.kt
@@ -1,6 +1,8 @@
-package org.wordpress.android.ui.stats.refresh.lists.widget
+package org.wordpress.android.ui.stats.refresh.lists.widget.today
 
 import org.wordpress.android.modules.AppComponent
+import org.wordpress.android.ui.stats.refresh.lists.widget.StatsWidget
+import org.wordpress.android.ui.stats.refresh.lists.widget.WidgetUpdater
 import javax.inject.Inject
 
 class StatsTodayWidget : StatsWidget() {
