+++ WordPress/src/main/java/org/wordpress/android/ui/stats/refresh/lists/widget/utils/StatsIntentUtils.kt
@@ -1,8 +1,8 @@
-package org.wordpress.android.ui.stats.refresh.lists.widget
+package org.wordpress.android.ui.stats.refresh.lists.widget.utils
 
 import android.content.Intent
-import org.wordpress.android.ui.stats.refresh.lists.widget.StatsWidgetConfigureFragment.ViewType
-import org.wordpress.android.ui.stats.refresh.lists.widget.StatsWidgetConfigureViewModel.Color
+import org.wordpress.android.ui.stats.refresh.lists.widget.configuration.StatsWidgetConfigureFragment.ViewType
+import org.wordpress.android.ui.stats.refresh.lists.widget.configuration.StatsWidgetConfigureViewModel.Color
 
 private const val COLOR_MODE_KEY = "color_mode_key"
 private const val VIEW_TYPE_KEY = "view_type_key"
