+++ WordPress/src/main/java/org/wordpress/android/ui/stats/refresh/lists/widget/views/StatsViewsWidget.kt
@@ -1,6 +1,8 @@
-package org.wordpress.android.ui.stats.refresh.lists.widget
+package org.wordpress.android.ui.stats.refresh.lists.widget.views
 
 import org.wordpress.android.modules.AppComponent
+import org.wordpress.android.ui.stats.refresh.lists.widget.StatsWidget
+import org.wordpress.android.ui.stats.refresh.lists.widget.WidgetUpdater
 import javax.inject.Inject
 
 class StatsViewsWidget : StatsWidget() {
