+++ WordPress/src/main/res/layout/stats_all_time_widget_configure_activity.xml
@@ -13,12 +13,12 @@
 
         <fragment
             android:id="@+id/fragment_container"
-            android:name="org.wordpress.android.ui.stats.refresh.lists.widget.StatsWidgetConfigureFragment"
+            android:name="org.wordpress.android.ui.stats.refresh.lists.widget.configuration.StatsWidgetConfigureFragment"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:layout_below="@id/toolbar"
-            app:viewType="allTime"
-            app:layout_behavior="@string/appbar_scrolling_view_behavior"/>
+            app:layout_behavior="@string/appbar_scrolling_view_behavior"
+            app:viewType="allTime"/>
 
         <com.google.android.material.appbar.AppBarLayout
             android:id="@+id/app_bar_layout"
