+++ WordPress/src/main/res/layout/stats_views_widget_configure_activity.xml
@@ -13,7 +13,7 @@
 
         <fragment
             android:id="@+id/fragment_container"
-            android:name="org.wordpress.android.ui.stats.refresh.lists.widget.StatsWidgetConfigureFragment"
+            android:name="org.wordpress.android.ui.stats.refresh.lists.widget.configuration.StatsWidgetConfigureFragment"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:layout_below="@id/toolbar"
