+++ WordPress/src/main/res/xml/stats_views_widget_info.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <appwidget-provider
     xmlns:android="http://schemas.android.com/apk/res/android"
-    android:configure="org.wordpress.android.ui.stats.refresh.lists.widget.StatsViewsWidgetConfigureActivity"
+    android:configure="org.wordpress.android.ui.stats.refresh.lists.widget.views.StatsViewsWidgetConfigureActivity"
     android:initialKeyguardLayout="@layout/stats_widget_list_light"
     android:initialLayout="@layout/stats_widget_list_light"
     android:minHeight="80dp"
